now this should all be named correctly
repos to repositories, revature to earth
