+++ DownToEarth/src/main/java/com/earth/DownToEarth/controllers/CommentController.java
@@ -1,0 +1,4 @@
+package com.earth.DownToEarth.controllers;
+
+public class CommentController {
+}
