+++ DownToEarth/src/main/java/com/earth/DownToEarth/models/Like.java
@@ -1,4 +1,4 @@
-package com.revature.DownToEarth.models;
+package com.earth.DownToEarth.models;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.*;
