+++ DownToEarth/src/main/java/com/earth/DownToEarth/models/User.java
@@ -1,11 +1,8 @@
-package com.revature.DownToEarth.models;
+package com.earth.DownToEarth.models;
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.*;
 
 import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.List;
 
 @AllArgsConstructor
 @NoArgsConstructor
