+++ DownToEarth/src/main/java/com/earth/DownToEarth/repositories/PostRepoImpl.java
@@ -1,6 +1,6 @@
-package com.revature.DownToEarth.repos;
+package com.earth.DownToEarth.repositories;
 
-import com.revature.DownToEarth.models.Post;
+import com.earth.DownToEarth.models.Post;
 import org.hibernate.Session;
 import org.springframework.stereotype.Repository;
