+++ DownToEarth/src/main/java/com/earth/DownToEarth/repositories/UserRepoImpl.java
@@ -1,12 +1,11 @@
-package com.revature.DownToEarth.repos;
+package com.earth.DownToEarth.repositories;
 
-import com.revature.DownToEarth.models.User;
+import com.earth.DownToEarth.models.User;
 import org.hibernate.Session;
 import org.springframework.stereotype.Repository;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
-import java.util.List;
 
 @Repository
 public class UserRepoImpl implements UserRepo{
