+++ DownToEarth/src/main/java/com/earth/DownToEarth/services/CommentService.java
@@ -1,0 +1,4 @@
+package com.earth.DownToEarth.services;
+
+public class CommentService {
+}
