--- DownToEarth/src/main/java/com/revature/DownToEarth/controllers/CommentController.java
@@ -1,4 +1,0 @@
-package com.revature.DownToEarth.controllers;
-
-public class CommentController {
-}
