--- DownToEarth/src/main/java/com/revature/DownToEarth/services/CommentService.java
@@ -1,4 +1,0 @@
-package com.revature.DownToEarth.services;
-
-public class CommentService {
-}
