--- DownToEarth/src/test/java/com/revature/DownToEarth/DownToEarthApplicationTests.java
@@ -1,13 +1,0 @@
-package com.revature.DownToEarth;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DownToEarthApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
